IT FINALLY WORKED
i was using a callback function in the forEach loop
and callback uses the global object (window) as its 'this'
i was expecting it to be my person{} object

the solution is to use an arrow function ()=>{}
bcs arrow function refer to the parent function 'this'
see more info from my research:

this.people.forEach(function (item) {
    // inside here, 'this' does NOT refer to your `people` object
    console.log(this.ulEl);  // undefined
});

The function passed to forEach is a regular function, which means its this value depends on how the function is called.

Since forEach calls the callback as a plain function, this inside that callback defaults to:

undefined in strict mode, or

the global object (window in browsers) in non-strict mode.

It does not automatically inherit the this from the outer method render.

How to fix it?
You want the callback to see the same this as the render method, i.e., the people object.

There are several common ways to achieve this:

1. Use arrow function (preferred)
Arrow functions don’t have their own this but inherit it from the surrounding lexical scope:

js

this.people.forEach((item) => {
    console.log(this.ulEl);  // Now works as expected
});

2. Capture this in a variable (self or that)
js

const self = this;
this.people.forEach(function (item) {
    console.log(self.ulEl);
});

3. Pass this as the second argument to forEach
js

this.people.forEach(function (item) {
    console.log(this.ulEl);
}, this);